testmain and http clean up
